drivers: rosy: touchscreen: Disable ESD check
While the ESD check is running, the touchscreen cannot be used. It
doesn't respond to interrupt either, in fact, it isn't even programmed
to; once the ESD check is started - nothing really can stop it. The only
way to make the touchscreen work, is to power-off-and-on the display,
essentialy forcing the touchscreen to detect panel reset; or, of course,
you can wait for the sudden ESD check to finish... but no one wants
that.

All of this is pitiful, so we disable the ESD check.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>